+++ dataiter/data_frame.py
@@ -716,7 +716,7 @@ class DataFrame(dict):
                 if dir not in [1, -1]:
                     raise ValueError("dir should be 1 or -1")
                 column = self[colname]
-                if dir < 0 and not (column.is_boolean or column.is_number):
+                if dir < 0 and not (column.is_boolean() or column.is_number()):
                     # Use rank for non-numeric types so that we can sort descending.
                     column = column.rank()
                 yield column if dir >= 0 else -column