+++ dataiter/test/test_vector.py
@@ -53,7 +53,7 @@ class TestVector:
         # Missing values should be None.
         a = Vector([True, False, np.nan, None])
         b = Vector([True, False, None, None])
-        assert a.is_object
+        assert a.is_object()
         assert a.equal(b)
 
     def test___new___missing_date(self):
@@ -61,7 +61,7 @@ class TestVector:
         # Missing values should be NaT.
         a = Vector([DATE, NaT, np.nan, None])
         b = Vector([DATE, NaT, NaT, NaT])
-        assert a.is_datetime
+        assert a.is_datetime()
         assert a.equal(b)
 
     def test___new___missing_datetime(self):
@@ -69,14 +69,14 @@ class TestVector:
         # Missing values should be NaT.
         a = Vector([DATETIME, np.nan, NaT, None])
         b = Vector([DATETIME, NaT, NaT, NaT])
-        assert a.is_datetime
+        assert a.is_datetime()
         assert a.equal(b)
 
     def test___new___missing_float(self):
         # Missing values should be NaN.
         a = Vector([1.1, 2.2, np.nan, None])
         b = Vector([1.1, 2.2, np.nan, np.nan])
-        assert a.is_float
+        assert a.is_float()
         assert a.equal(b)
 
     def test___new___missing_integer(self):
@@ -84,14 +84,14 @@ class TestVector:
         # Missing values should be NaN.
         a = Vector([1, 2, np.nan, None])
         b = Vector([1, 2, np.nan, np.nan])
-        assert a.is_float
+        assert a.is_float()
         assert a.equal(b)
 
     def test___new___missing_string(self):
         # Missing values should be blank strings.
         a = Vector(["a", "b", "", np.nan, None])
         b = Vector(["a", "b", "", "", ""])
-        assert a.is_string
+        assert a.is_string()
         assert a.equal(b)
 
     def test__array_wrap___expect_vector(self):
@@ -109,47 +109,47 @@ class TestVector:
 
     def test_as_boolean(self):
         a = Vector([1, 0]).as_boolean()
-        assert a.is_boolean
+        assert a.is_boolean()
         assert np.all(a == [True, False])
 
     def test_as_bytes(self):
         a = Vector([0, 1]).as_bytes()
-        assert a.is_bytes
+        assert a.is_bytes()
         assert np.all(a == [b"0", b"1"])
 
     def test_as_bytes_string(self):
         a = Vector(["a", "ö"]).as_bytes()
-        assert a.is_bytes
+        assert a.is_bytes()
         assert np.all(a == [b"a", b"\xc3\xb6"])
 
     def test_as_date(self):
         a = Vector([DATETIME]).as_date()
-        assert a.is_datetime
+        assert a.is_datetime()
         assert np.all(a == [np.datetime64(DATE, "D")])
 
     def test_as_datetime(self):
         a = Vector([DATE]).as_datetime()
-        assert a.is_datetime
+        assert a.is_datetime()
         assert np.all(a == [np.datetime64(DATE, "us")])
 
     def test_as_float(self):
         a = Vector([1, 2]).as_float()
-        assert a.is_float
+        assert a.is_float()
         assert np.all(a == [1.0, 2.0])
 
     def test_as_integer(self):
         a = Vector([1.1, 2.2]).as_integer()
-        assert a.is_integer
+        assert a.is_integer()
         assert np.all(a == [1, 2])
 
     def test_as_object(self):
         a = Vector([1, 2]).as_object()
-        assert a.is_object
+        assert a.is_object()
         assert np.all(a == [1, 2])
 
     def test_as_string(self):
         a = Vector([1, 2]).as_string()
-        assert a.is_string
+        assert a.is_string()
         assert np.all(a == ["1", "2"])
 
     def test_equal_boolean(self):
@@ -191,7 +191,7 @@ class TestVector:
     def test_fast(self):
         a = Vector.fast([1, 2, 3], int)
         b = Vector([1, 2, 3], int)
-        assert a.is_integer
+        assert a.is_integer()
         assert a.equal(b)
 
     def test_fast_iterator(self):
@@ -208,54 +208,54 @@ class TestVector:
         assert a.head(3).tolist() == [1, 2, 3]
 
     def test_is_boolean(self):
-        assert not Vector([b"1"]).is_boolean
-        assert Vector([True]).is_boolean
-        assert not Vector([1]).is_boolean
-        assert not Vector([1.1]).is_boolean
-        assert not Vector(["a"]).is_boolean
-        assert not Vector([DATE]).is_boolean
-        assert not Vector([DATETIME]).is_boolean
-        assert not Vector([self]).is_boolean
+        assert not Vector([b"1"]).is_boolean()
+        assert Vector([True]).is_boolean()
+        assert not Vector([1]).is_boolean()
+        assert not Vector([1.1]).is_boolean()
+        assert not Vector(["a"]).is_boolean()
+        assert not Vector([DATE]).is_boolean()
+        assert not Vector([DATETIME]).is_boolean()
+        assert not Vector([self]).is_boolean()
 
     def test_is_bytes(self):
-        assert Vector([b"1"]).is_bytes
-        assert not Vector([True]).is_bytes
-        assert not Vector([1]).is_bytes
-        assert not Vector([1.1]).is_bytes
-        assert not Vector(["a"]).is_bytes
-        assert not Vector([DATE]).is_bytes
-        assert not Vector([DATETIME]).is_bytes
-        assert not Vector([self]).is_bytes
+        assert Vector([b"1"]).is_bytes()
+        assert not Vector([True]).is_bytes()
+        assert not Vector([1]).is_bytes()
+        assert not Vector([1.1]).is_bytes()
+        assert not Vector(["a"]).is_bytes()
+        assert not Vector([DATE]).is_bytes()
+        assert not Vector([DATETIME]).is_bytes()
+        assert not Vector([self]).is_bytes()
 
     def test_is_datetime(self):
-        assert not Vector([b"1"]).is_datetime
-        assert not Vector([True]).is_datetime
-        assert not Vector([1]).is_datetime
-        assert not Vector([1.1]).is_datetime
-        assert not Vector(["a"]).is_datetime
-        assert Vector([DATE]).is_datetime
-        assert Vector([DATETIME]).is_datetime
-        assert not Vector([self]).is_datetime
+        assert not Vector([b"1"]).is_datetime()
+        assert not Vector([True]).is_datetime()
+        assert not Vector([1]).is_datetime()
+        assert not Vector([1.1]).is_datetime()
+        assert not Vector(["a"]).is_datetime()
+        assert Vector([DATE]).is_datetime()
+        assert Vector([DATETIME]).is_datetime()
+        assert not Vector([self]).is_datetime()
 
     def test_is_float(self):
-        assert not Vector([b"1"]).is_float
-        assert not Vector([True]).is_float
-        assert not Vector([1]).is_float
-        assert Vector([1.1]).is_float
-        assert not Vector(["a"]).is_float
-        assert not Vector([DATE]).is_float
-        assert not Vector([DATETIME]).is_float
-        assert not Vector([self]).is_float
+        assert not Vector([b"1"]).is_float()
+        assert not Vector([True]).is_float()
+        assert not Vector([1]).is_float()
+        assert Vector([1.1]).is_float()
+        assert not Vector(["a"]).is_float()
+        assert not Vector([DATE]).is_float()
+        assert not Vector([DATETIME]).is_float()
+        assert not Vector([self]).is_float()
 
     def test_is_integer(self):
-        assert not Vector([b"1"]).is_integer
-        assert not Vector([True]).is_integer
-        assert Vector([1]).is_integer
-        assert not Vector([1.1]).is_integer
-        assert not Vector(["a"]).is_integer
-        assert not Vector([DATE]).is_integer
-        assert not Vector([DATETIME]).is_integer
-        assert not Vector([self]).is_integer
+        assert not Vector([b"1"]).is_integer()
+        assert not Vector([True]).is_integer()
+        assert Vector([1]).is_integer()
+        assert not Vector([1.1]).is_integer()
+        assert not Vector(["a"]).is_integer()
+        assert not Vector([DATE]).is_integer()
+        assert not Vector([DATETIME]).is_integer()
+        assert not Vector([self]).is_integer()
 
     def test_is_missing_date(self):
         a = Vector([DATE, DATE, NaT])
@@ -278,34 +278,34 @@ class TestVector:
         assert a.is_missing().tolist() == [False, False, True]
 
     def test_is_number(self):
-        assert not Vector([b"1"]).is_number
-        assert not Vector([True]).is_number
-        assert Vector([1]).is_number
-        assert Vector([1.1]).is_number
-        assert not Vector(["a"]).is_number
-        assert not Vector([DATE]).is_number
-        assert not Vector([DATETIME]).is_number
-        assert not Vector([self]).is_number
+        assert not Vector([b"1"]).is_number()
+        assert not Vector([True]).is_number()
+        assert Vector([1]).is_number()
+        assert Vector([1.1]).is_number()
+        assert not Vector(["a"]).is_number()
+        assert not Vector([DATE]).is_number()
+        assert not Vector([DATETIME]).is_number()
+        assert not Vector([self]).is_number()
 
     def test_is_object(self):
-        assert not Vector([b"1"]).is_object
-        assert not Vector([True]).is_object
-        assert not Vector([1]).is_object
-        assert not Vector([1.1]).is_object
-        assert not Vector(["a"]).is_object
-        assert not Vector([DATE]).is_object
-        assert not Vector([DATETIME]).is_object
-        assert Vector([self]).is_object
+        assert not Vector([b"1"]).is_object()
+        assert not Vector([True]).is_object()
+        assert not Vector([1]).is_object()
+        assert not Vector([1.1]).is_object()
+        assert not Vector(["a"]).is_object()
+        assert not Vector([DATE]).is_object()
+        assert not Vector([DATETIME]).is_object()
+        assert Vector([self]).is_object()
 
     def test_is_string(self):
-        assert not Vector([b"1"]).is_string
-        assert not Vector([True]).is_string
-        assert not Vector([1]).is_string
-        assert not Vector([1.1]).is_string
-        assert Vector(["a"]).is_string
-        assert not Vector([DATE]).is_string
-        assert not Vector([DATETIME]).is_string
-        assert not Vector([self]).is_string
+        assert not Vector([b"1"]).is_string()
+        assert not Vector([True]).is_string()
+        assert not Vector([1]).is_string()
+        assert not Vector([1.1]).is_string()
+        assert Vector(["a"]).is_string()
+        assert not Vector([DATE]).is_string()
+        assert not Vector([DATETIME]).is_string()
+        assert not Vector([self]).is_string()
 
     def test_map(self):
         a = Vector([1, 2, 3, 4, 5])